requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
SQLAlchemy==2.0.23
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
itsdangerous==2.1.2
python-dotenv==1.0.0
schedule==1.2.1
Pillow==10.1.0
GitPython==3.1.40
requests==2.31.0 